# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ViewCount {
  /** @minimum 0 */
  count: number;
}

export interface ViewsLoginInput {
  /** @minLength 1 */
  password: string;
}

export interface ViewsAdminSession {
  authenticated: boolean;
}

export interface ViewRecord {
  id: number;
  ip: string;
  timestamp: string;
  userAgent: string;
}

export interface ViewsHistory {
  /** @minimum 0 */
  count: number;
  visits: ViewRecord[];
}