# lib/api-zod/src/generated/types/viewRecord.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ViewRecord {
  id: number;
  ip: string;
  timestamp: Date;
  userAgent: string;
}