# lib/api-zod/src/generated/types/viewsHistory.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ViewRecord } from './viewRecord';

export interface ViewsHistory {
  /** @minimum 0 */
  count: number;
  visits: ViewRecord[];
}